requirements.txt
torch
numpy
pynvml; platform_machine != 'aarch64'
py-cpuinfo
jetson-stats; platform_machine == 'aarch64'
